+++ src/backend_raqote/clip_and_mask.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use crate::{prelude::*, ConvTransform};
+use crate::{prelude::*, ConvTransform, RenderState};
 use super::{path, RaqoteLayers, RaqoteDrawTargetExt};
 
 
@@ -142,7 +142,7 @@ pub fn mask(
             mask_dt.transform(&usvg::Transform::from_bbox(bbox).to_native());
         }
 
-        super::render_group(node, opt, layers, &mut mask_dt);
+        super::render_group(node, opt, &mut RenderState::Ok, layers, &mut mask_dt);
         mask_dt.pop_clip();
     }