+++ src/filter.rs
@@ -124,9 +124,10 @@ pub trait Filter<T: ImageExt> {
         bbox: Option<Rect>,
         ts: &usvg::Transform,
         opt: &Options,
+        background: Option<&T>,
         canvas: &mut T,
     ) {
-        let res = Self::_apply(filter, bbox, ts, opt, canvas);
+        let res = Self::_apply(filter, bbox, ts, opt, background, canvas);
 
         // Clear on error.
         if res.is_err() {
@@ -152,6 +153,7 @@ pub trait Filter<T: ImageExt> {
         bbox: Option<Rect>,
         ts: &usvg::Transform,
         opt: &Options,
+        background: Option<&T>,
         canvas: &mut T,
     ) -> Result<(), Error> {
         let mut results = Vec::new();
@@ -165,42 +167,42 @@ pub trait Filter<T: ImageExt> {
 
             let mut result = match primitive.kind {
                 usvg::FilterKind::FeBlend(ref fe) => {
-                    let input1 = Self::get_input(&fe.input1, region, &results, canvas)?;
-                    let input2 = Self::get_input(&fe.input2, region, &results, canvas)?;
+                    let input1 = Self::get_input(&fe.input1, region, &results, background, canvas)?;
+                    let input2 = Self::get_input(&fe.input2, region, &results, background, canvas)?;
                     Self::apply_blend(fe, cs, region, input1, input2)
                 }
                 usvg::FilterKind::FeFlood(ref fe) => {
                     Self::apply_flood(fe, region)
                 }
                 usvg::FilterKind::FeGaussianBlur(ref fe) => {
-                    let input = Self::get_input(&fe.input, region, &results, canvas)?;
+                    let input = Self::get_input(&fe.input, region, &results, background, canvas)?;
                     Self::apply_blur(fe, filter.primitive_units, cs, bbox, ts, input)
                 }
                 usvg::FilterKind::FeOffset(ref fe) => {
-                    let input = Self::get_input(&fe.input, region, &results, canvas)?;
+                    let input = Self::get_input(&fe.input, region, &results, background, canvas)?;
                     Self::apply_offset(fe, filter.primitive_units, bbox, ts, input)
                 }
                 usvg::FilterKind::FeComposite(ref fe) => {
-                    let input1 = Self::get_input(&fe.input1, region, &results, canvas)?;
-                    let input2 = Self::get_input(&fe.input2, region, &results, canvas)?;
+                    let input1 = Self::get_input(&fe.input1, region, &results, background, canvas)?;
+                    let input2 = Self::get_input(&fe.input2, region, &results, background, canvas)?;
                     Self::apply_composite(fe, cs, region, input1, input2)
                 }
                 usvg::FilterKind::FeMerge(ref fe) => {
-                    Self::apply_merge(fe, cs, region, &results, canvas)
+                    Self::apply_merge(fe, cs, region, &results, background, canvas)
                 }
                 usvg::FilterKind::FeTile(ref fe) => {
-                    let input = Self::get_input(&fe.input, region, &results, canvas)?;
+                    let input = Self::get_input(&fe.input, region, &results, background, canvas)?;
                     Self::apply_tile(input, region)
                 }
                 usvg::FilterKind::FeImage(ref fe) => {
                     Self::apply_image(fe, region, subregion, opt)
                 }
                 usvg::FilterKind::FeComponentTransfer(ref fe) => {
-                    let input = Self::get_input(&fe.input, region, &results, canvas)?;
+                    let input = Self::get_input(&fe.input, region, &results, background, canvas)?;
                     Self::apply_component_transfer(fe, cs, input)
                 }
                 usvg::FilterKind::FeColorMatrix(ref fe) => {
-                    let input = Self::get_input(&fe.input, region, &results, canvas)?;
+                    let input = Self::get_input(&fe.input, region, &results, background, canvas)?;
                     Self::apply_color_matrix(fe, cs, input)
                 }
             }?;
@@ -244,6 +246,7 @@ pub trait Filter<T: ImageExt> {
         input: &usvg::FilterInput,
         region: ScreenRect,
         results: &[FilterResult<T>],
+        background: Option<&T>,
         canvas: &T,
     ) -> Result<Image<T>, Error>;
 
@@ -285,6 +288,7 @@ pub trait Filter<T: ImageExt> {
         cs: ColorSpace,
         region: ScreenRect,
         results: &[FilterResult<T>],
+        background: Option<&T>,
         canvas: &T,
     ) -> Result<Image<T>, Error>;