+++ usvg/src/convert/marker.rs
@@ -156,12 +156,9 @@ fn resolve(
 
         // TODO: do not create a group when no clipPath
         let mut g_node = parent.append_kind(tree::NodeKind::Group(tree::Group {
-            id: String::new(),
             transform: ts,
-            opacity: tree::Opacity::default(),
             clip_path: clip_path.clone(),
-            mask: None,
-            filter: None,
+            .. tree::Group::default()
         }));
 
         let mut marker_state = state.clone();